incidentd: fix data size formatting on LoongArch64
Cast the on-disk file size to int64_t before formatting it with PRIi64. LoongArch64 exposes st_size as long long while int64_t is long, and the mismatch is rejected under -Werror=format.

+++ cmds/incidentd/src/WorkDirectory.cpp
@@ -416,7 +416,8 @@ status_t ReportFile::startFilteringData(int writeFd, const IncidentReportArgs& a
     if (st.st_size != mEnvelope.data_file_size()) {
         ALOGW("File size mismatch. Envelope says %" PRIi64 " bytes but data file is %" PRIi64
               " bytes: %s",
-              (int64_t)mEnvelope.data_file_size(), st.st_size, mDataFileName.c_str());
+              (int64_t)mEnvelope.data_file_size(), static_cast<int64_t>(st.st_size),
+              mDataFileName.c_str());
         ALOGW("Removing incident report");
         mWorkDirectory->remove(this);
         close(writeFd);
@@ -872,4 +873,3 @@ void get_args_from_report(IncidentReportArgs* out, const ReportFileProto_Report&
 }  // namespace incidentd
 }  // namespace os
 }  // namespace android
-
